requirements.txt
Django==1.9.7
certifi==2017.4.17
chardet==3.0.4
dj-database-url==0.4.1
django-watson==1.3.1
gunicorn==19.6.0
idna==2.5
pprint==0.1
psycopg2==2.6.2
requests==2.18.1
urllib3==1.21.1
whitenoise==2.0.6
